refactoring/cleanup
